refactor(orchestrator): extract approval lifecycle from runner.go
After the queue + reconcile cluster moved out, runner.go still
held the approval-resolved + approval-rejected flow plus the
per-task approval spec/create helpers. Pull that cluster into a
sibling so the next reader sees "lifecycle + run events + I/O"
in runner.go and "what happens when an approval resolves" in
runner_approvals.go.

  runner.go:            1829 → 1574 LOC
  runner_approvals.go:  new, 260 LOC

What moves to runner_approvals.go

  Approval-resolved transitions
  - ResumeTaskAfterApproval: resumes a task whose approval was
    granted, building the new run + step events that follow.
  - RejectTaskAfterApproval: terminates a task whose approval was
    rejected as failed_at_approval, emitting the rejection events.

  Cleanup + telemetry
  - cancelPendingApprovalsForRun: marks any other pending
    approvals on the same run as cancelled when the run resolves
    by a different path (timeout, cancel, downstream failure).
  - recordApprovalResolved: trace.Record helper that emits the
    standard approval.resolved telemetry event with the wait-time
    metric.

  Per-task approval spec
  - approvalRequiredForTask, approvalSpecForTask: decide whether a
    task needs an approval and what kind/reason text to display.
  - createApprovalForTask: builds + persists the TaskApproval row
    and emits the approval.requested event.

What stays in runner.go

  Lifecycle (StartTask, ResumeTask, RetryTask, ContinueTask,
  CancelTask and the *FromTurn variants), event emission
  (emitRunEvent), step/artifact upsert (upsertStep, upsertArtifact,
  gitSummaryArtifact, resumeCheckpointForRun, finalizeFailedRun),
  Runner construction. Events + IO clusters split out in 4c.

The receiver type stays *Runner; no public API change; no test
changes. Imports per file pruned via goimports.

Verification

  go build ./...                                — clean
  go vet ./...                                  — clean
  gofmt -s -l .                                 — clean
  go test ./internal/orchestrator
    -run 'TestRunner|runner_*' -race -count=5   — 50 / 50 pass
  go test ./... -race -count=1                  — 1779 / 1779 pass
